Button Delete Finish

+++ src/main/java/com/test/repository/UserRepository.java
@@ -1,11 +1,15 @@
 package com.test.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.test.model.User;
 
 @Repository("userRepository")
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> 
+{
 	 User findByEmail(String email);
+//	 void findOneby(Integer id);
 }
